+++ RELEASE_NOTES.md
@@ -22,6 +22,7 @@ More details at <a href="https://nighttime-imaging.eu/donate/" target="_blank">n
 - N.I.N.A. Star Detector now properly calculates the centroid of stars
   - The centroid is now also annotated in the N.I.N.A. Star Annotator
 - Autofocus report now displays the determined step number as an integer instead of a floating point number
+- N.I.N.A. will now query the ASCOM device name after connecting to the driver.
 
 ## Commandline Options
 The following command line options have been added